requirements.txt
requests==2.25.1
openai
beautifulsoup4
djangorestframework
pymysql